new folder Common, movie and people page moved
