+++ src/App.js
@@ -1,6 +1,6 @@
 import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
-import MoviePeople from "./features/MoviePeople";
-import MoviesList from "./features/MoviesList";
+import MoviePeople from "./features/Content/MoviePeople";
+import MoviesList from "./features/Content/MoviesList";
 import { StyledInput, StyledList, StyledNav, StyledNavLink, StyledSearch } from "./styled";
 
 function App() {
